Multiple fixes : - Changed default precision for currency (now 6) - Simplified LocaleDataSource dependencies (now top layer instead of an array of layers) - fixed currency data mapping in the CLDR xml reader (numeric ISO code must be a string)

+++ src/Core/Localization/CLDR/Reader.php
@@ -506,7 +506,7 @@ class Reader implements ReaderInterface
                 if (!empty($codesMapping)) {
                     /** @var SimplexmlElement $codesMapping */
                     $codesMapping                 = $codesMapping[0];
-                    $currencyData->numericIsoCode = (int)(string)$codesMapping->attributes()->numeric;
+                    $currencyData->numericIsoCode = (string)$codesMapping->attributes()->numeric;
                 }
 
                 $fractionsData = $this->supplementalXml->supplementalData->xpath(
